refactor: refactor controller
