+++ src/controller/api.ts
@@ -1,20 +1,17 @@
+import { Kv } from "../infrastructure/kv.ts";
+import { getRequest } from "../service/request.ts";
+import { ItemUseCase } from "../usecase/item/impl.ts";
+import { ItemRepository } from "../repository/item/impl.ts";
+
 import type { Context } from "hono";
 
-import { getRequest } from "../services/request.ts";
-import {
-  deleteItem,
-  disableItem,
-  getAllItems,
-  getItem,
-  updateItem,
-  upsertItem,
-} from "../services/items.ts";
+const itemUsecase = new ItemUseCase(new ItemRepository(await Kv.getKv()));
 
 export const getAllItemsController = async (c: Context) => {
   const { host } = getRequest(c);
   if (!host) return c.notFound();
 
-  const items = await getAllItems(host);
+  const items = await itemUsecase.findAllItems(host);
   return c.json(items);
 };
 
@@ -22,7 +19,7 @@ export const getItemController = async (c: Context) => {
   const { host, param } = getRequest(c);
   if (!host || !param) return c.notFound();
 
-  const item = await getItem(host, param);
+  const item = await itemUsecase.findItem(host, param);
   if (!item) return c.notFound();
   return c.json(item);
 };
@@ -36,15 +33,14 @@ export const putItemController = async (
   const { host, param } = getRequest(c);
   if (!host || !param) return c.notFound();
 
-  const item = typeof count === "number" ? null : await getItem(host, param);
-  const res = await upsertItem(host, param, {
+  const item = await itemUsecase.findItem(host, param);
+  const res = await itemUsecase.upsertItem(host, {
     param,
     description: description ?? item?.description,
     url,
     count: typeof count === "number" ? count : item?.count ?? 0,
     unavailable: false,
   });
-
   return c.json(res);
 };
 
@@ -58,10 +54,10 @@ export const patchItemController = async (
   const { host, param } = getRequest(c);
   if (!host || !param) return c.notFound();
 
-  const item = await getItem(host, param);
+  const item = await itemUsecase.findItem(host, param);
   if (!item) return c.notFound();
 
-  const res = await updateItem(host, param, {
+  const res = await itemUsecase.updateItem(host, item, {
     description,
     url,
     count,
@@ -74,11 +70,14 @@ export const deleteItemController = async (c: Context) => {
   const { host, param } = getRequest(c);
   if (!host || !param) return c.notFound();
 
+  const item = await itemUsecase.findItem(host, param);
+  if (!item) return c.notFound();
+
   if (c.req.query("permanently") === "true") {
-    await deleteItem(host, param);
+    await itemUsecase.deleteItem(host, item);
     return c.body(null, 204);
   }
 
-  const res = await disableItem(host, param);
+  const res = await itemUsecase.disableItem(host, item);
   return c.json(res);
 };
